requirements/base.txt
pyquery==1.2.4
requests>=0.12.0
selenium
PyYAML==3.10
lxml==2.2
